+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.47.0 - TBD
+## 0.47.0 - 2026-01-20
 
 ### Enhancements
 - Changed the `__repr__` implementation of all records in Python to be more Pythonic
@@ -10,6 +10,8 @@
   - `LowerPriceLimit`
   - `BlockVolume`
   - `VenueSpecificVolume1`
+- Upgraded `pyo3` version to 0.27.2
+- Upgraded `time` version to 0.3.45
 
 ## 0.46.0 - 2026-01-06